wip: tweak sizing on ani
