+++ src/App.tsx
@@ -1,7 +1,7 @@
-import { ContainerInner, DashboardContainer } from './theme/styles';
+import { BackgroundImage, ContainerInner, DashboardContainer } from './theme/styles';
 import { SideBar } from './containers/SideBar';
 import { Dashboard } from './containers/Dashboard';
-import { AppBackground } from './containers/AppBackground';
+
 import { useUIStore } from './store/useUIStore.ts';
 import { useGetStatus } from './hooks/useGetStatus.ts';
 import { useSetUp } from './hooks/useSetUp.ts';
@@ -29,13 +29,14 @@ export default function App() {
     const isShuttingDown = useShuttingDown();
     const showSplash = useUIStore((s) => s.showSplash);
     const view = useUIStore((s) => s.view);
+    const visualMode = useUIStore((s) => s.visualMode);
 
     const canRenderMain = !isShuttingDown && !showSplash;
     const splashScreenMarkup = <SplashScreen />;
     const shutDownMarkup = isShuttingDown ? <ShuttingDownScreen /> : null;
     const mainMarkup = canRenderMain ? (
-        <DashboardContainer layout>
-            <ContainerInner layout>
+        <DashboardContainer>
+            <ContainerInner>
                 <SideBar />
                 <AirdropLogin />
                 <Dashboard status={view} />
@@ -47,11 +48,11 @@ export default function App() {
         <ThemeProvider>
             <GlobalReset />
             <GlobalStyle />
+            <AutoUpdateDialog />
             <LayoutGroup>
-                <AppBackground />
-                <AutoUpdateDialog />
                 {splashScreenMarkup}
                 {shutDownMarkup}
+                {!visualMode || view != 'mining' ? <BackgroundImage layout /> : null}
                 {mainMarkup}
                 <ErrorSnackbar />
             </LayoutGroup>
