--- src/containers/AppBackground/index.ts
@@ -1,1 +1,0 @@
-export { default as AppBackground } from './AppBackground';
