+++ src/hooks/useSetUp.ts
@@ -1,4 +1,4 @@
-import { useEffect, useRef } from 'react';
+import { useLayoutEffect, useRef } from 'react';
 import { listen } from '@tauri-apps/api/event';
 import { TauriEvent } from '../types.ts';
 
@@ -15,21 +15,19 @@ export function useSetUp() {
     const startupInitiated = useRef(false);
     const setView = useUIStore((s) => s.setView);
     const setSetupDetails = useAppStateStore((s) => s.setSetupDetails);
-    const settingUpFinished = useAppStateStore(useShallow((s) => s.settingUpFinished));
     const setError = useAppStateStore((s) => s.setError);
 
     const isAfterAutoUpdate = useAppStateStore(useShallow((s) => s.isAfterAutoUpdate));
 
     useVersions();
 
-    useEffect(() => {
+    useLayoutEffect(() => {
         const unlistenPromise = listen('message', ({ event: e, payload: p }: TauriEvent) => {
             console.info('Setup Event:', e, p);
             switch (p.event_type) {
                 case 'setup_status':
                     setSetupDetails(p.title, p.title_params, p.progress);
                     if (p.progress >= 1) {
-                        settingUpFinished();
                         setAnimationState('showVisual');
                         setView('mining');
                     }
@@ -52,5 +50,5 @@ export function useSetUp() {
         return () => {
             unlistenPromise.then((unlisten) => unlisten());
         };
-    }, [setError, setSetupDetails, setView, settingUpFinished, isAfterAutoUpdate]);
+    }, [isAfterAutoUpdate, setError, setSetupDetails, setView]);
 }
